Smirkey Drive
Mr Smirkey the protagonist is wandering on the boulevards and
eventually Smirkey falls into a hole which he himself doesn't know what
it is , this moment the game begins. Now when when Mr. Smirkey falls
into the tunnel, it has to dodge some evil smirks in order to proceed.
The more the user dodges the greater the user scores.